Add files via upload
Given a real anonymized dataset from an insurance company and an external dataset, I was tasked with simulating the role of a Data Analysis team’s early work and give a presentation of said work to management. 
The goal is to integrate an external data package purchased by the insurance company to better understand the customer and contribute materially to operations by creating a data product. 
To prepare the data for processing in Python, I cleaned, transformed, and merged the company and external data.
I used Python to create machine learning models to gain insight on different customer segments to target with marketing, and finally create a model attempting to predict if a policy would have a high payout.